Move GlobalCss to own file in client assets
