+++ client/src/utilities/theme.tsx
@@ -1,4 +1,4 @@
-import { createMuiTheme, withStyles } from '@material-ui/core/styles';
+import { createMuiTheme } from '@material-ui/core/styles';
 import { ThemeOptions } from '@material-ui/core/styles/createMuiTheme';
 
 
@@ -7,11 +7,3 @@ export const theme = (props: string): ThemeOptions => createMuiTheme({
     type: props === 'dark' ? 'dark' : 'light',
   },
 });
-
-export const GlobalCss = withStyles({
-  '@global': {
-    '.MuiPaper-root': {
-      borderRadius: 0,
-    },
-  },
-})(() => null);
